Implement executable segment in code directory
Previous attempt to use version 0x20400 failed because the CD structure was
not updated. This implements the executable segment info and works on iOS
14.2.

Additionally, the execSegFlags is set to `CS_EXECSEG_ALLOW_UNSIGNED` for the
main executable whenever the entitlement `get-task-allow` is seen. This
allows for bulletproof JIT to work on iOS 14.2.

Currently, the flag is set whenever `get-task-allow` is seen but a proper
implementation should parse the entitlement and check that the key is set
to true before setting the flag.